# docs/swagger.yaml
basePath: /api/v1
definitions:
  handler.errorResponse:
    properties:
      error:
        type: string
    type: object
...
  model.GenerateSalesReportInput:
    properties:
      from:
        type: string
      title:
        maxLength: 150
        type: string
      to:
        type: string
    type: object
...
  model.Product:
    properties:
      category_id:
        type: integer
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      price:
        type: number
      supplier_id:
        type: integer
      updated_at:
        type: string
    type: object
  model.ProductInput:
    properties:
      category_id:
        type: integer
      description:
        type: string
      name:
        maxLength: 100
        type: string
      price:
        minimum: 0
        type: number
      supplier_id:
        type: integer
    required:
    - name
    type: object
  model.PurchaseRequest:
    properties:
      comment:
        type: string
      created_at:
        type: string
      id:
        type: integer
      product_id:
        type: integer
      quantity:
        type: integer
      status:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  model.PurchaseRequestInput:
    properties:
      comment:
        type: string
      product_id:
        type: integer
      quantity:
        minimum: 1
        type: integer
    required:
    - product_id
    - quantity
    type: object
  model.PurchaseStatusInput:
    properties:
      status:
        enum:
        - new
        - in_progress
        - approved
        - rejected
        - completed
        type: string
    required:
    - status
    type: object
...
  model.Report:
    properties:
      category_id:
        type: integer
      content:
        type: string
      created_at:
        type: string
      created_date:
        type: string
      id:
        type: integer
      title:
        type: string
      updated_at:
        type: string
    type: object
  model.ReportInput:
    properties:
      category_id:
        type: integer
      content:
        type: string
      title:
        maxLength: 150
        type: string
    required:
    - title
    type: object
...
host: localhost:8080
info:
  contact:
    name: CRM Backend
  description: |-
    Бэкенд CRM-системы для предприятия оптово-розничной торговли:
    управление продуктами, категориями, поставщиками, заявками на закупку,
    счетами-фактурами, аналитикой и отчётами.
  license:
    name: MIT
  termsOfService: http://swagger.io/terms/
  title: CRM API
  version: "1.0"
paths:
...
  /products:
    get:
      parameters:
      - description: Фильтр по категории
        in: query
        name: category_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Product'
            type: array
      security:
      - BearerAuth: []
      summary: Список продуктов
      tags:
      - products
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные продукта
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.ProductInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Создать продукт
      tags:
      - products
  /products/{id}:
    delete:
      parameters:
      - description: ID продукта
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Удалить продукт
      tags:
      - products
    get:
      parameters:
      - description: ID продукта
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Product'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Получить продукт
      tags:
      - products
    put:
      consumes:
      - application/json
      parameters:
      - description: ID продукта
        in: path
        name: id
        required: true
        type: integer
      - description: Данные продукта
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.ProductInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Product'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Обновить продукт
      tags:
      - products
  /purchase-requests:
    get:
      description: Администратор видит все заявки, пользователь — только свои
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.PurchaseRequest'
            type: array
      security:
      - BearerAuth: []
      summary: Список заявок на закупку
      tags:
      - purchase-requests
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные заявки
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.PurchaseRequestInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.PurchaseRequest'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Создать заявку на закупку
      tags:
      - purchase-requests
  /purchase-requests/{id}:
    delete:
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Удалить заявку
      tags:
      - purchase-requests
    get:
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.PurchaseRequest'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Получить заявку
      tags:
      - purchase-requests
  /purchase-requests/{id}/status:
    patch:
      consumes:
      - application/json
      description: Доступно только администратору
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      - description: Новый статус
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.PurchaseStatusInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.PurchaseRequest'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Сменить статус заявки
      tags:
      - purchase-requests
  /reports:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Report'
            type: array
      security:
      - BearerAuth: []
      summary: Список отчётов
      tags:
      - reports
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные отчёта
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.ReportInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Report'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Создать отчёт вручную
      tags:
      - reports
  /reports/{id}:
    delete:
      parameters:
      - description: ID отчёта
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Удалить отчёт
      tags:
      - reports
    get:
      parameters:
      - description: ID отчёта
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Report'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Получить отчёт
      tags:
      - reports
  /reports/{id}/export:
    get:
      description: Отдаёт содержимое отчёта как CSV-файл (UTF-8 BOM, для Excel)
      parameters:
      - description: ID отчёта
        in: path
        name: id
        required: true
        type: integer
      produces:
      - text/csv
      responses:
        "200":
          description: CSV-файл
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Экспорт отчёта в CSV
      tags:
      - reports
  /reports/generate-sales:
    post:
      consumes:
      - application/json
      description: Генерирует отчёт за период на основе реальной аналитики и сохраняет
        его (содержимое — CSV)
      parameters:
      - description: Период (поля опциональны)
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.GenerateSalesReportInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Report'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Сформировать отчёт о продажах
      tags:
      - reports
...
securityDefinitions:
  BearerAuth:
    description: 'Введите токен в формате: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...